+++ lesson_4/testDir/text.txt
@@ -1,0 +1,1 @@
+Test
